+++ src/renderer/port-forward/port-forward-utils.ts
@@ -19,7 +19,36 @@
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-export * from "./port-forward.store";
-export * from "./port-forward-item";
-export * from "./port-forward-dialog";
-export * from "./port-forward-utils";
+
+import { openExternal } from "../utils";
+import { Notifications } from "../components/notifications";
+import type { ForwardedPort } from "./port-forward-item";
+import logger from "../../common/logger";
+
+export function portForwardAddress(portForward: ForwardedPort) {
+    return `${portForward.protocol ?? "http"}://localhost:${portForward.forwardPort}`;
+  }
+  
+  export function openPortForward(portForward: ForwardedPort) {
+    const browseTo = portForwardAddress(portForward);
+  
+    openExternal(browseTo)
+      .catch(error => {
+        logger.error(`failed to open in browser: ${error}`, {
+          clusterId: portForward.clusterId,
+          port: portForward.port,
+          kind: portForward.kind,
+          namespace: portForward.namespace,
+          name: portForward.name,
+        });
+        Notifications.error(`Failed to open ${browseTo} in browser`);
+      }
+      );
+  
+  }
+  
+  export function predictProtocol(name: string) {
+    return name === "https" ? "https" : "http";
+  }
+  
+
